requirements.txt
scikit-learn
joblib
flask
pillow
numpy
